factor out getImageUrls
